+++ ebpub/ebpub/db/tests/test_models.py
@@ -215,3 +215,17 @@ class DatabaseExtensionsTestCase(TestCase):
         self.assertEqual(top_lookups[1]['count'], 2)
         self.assertEqual(top_lookups[1]['lookup'].slug, u'tag-2')
 
+    def test_allowed_schema_ids(self):
+        from ebpub.db.models import Schema
+        self.assertEqual(Schema.objects.allowed_schema_ids(),
+                         [1])
+        self.assertEqual(Schema.public_objects.allowed_schema_ids(),
+                         [1])
+        Schema.objects.all().update(is_public=False)
+        # Now 'objects' still returns everything, but 'public_objects' doesn't.
+        self.assertEqual(Schema.objects.allowed_schema_ids(),
+                         [1])
+        self.assertEqual(Schema.public_objects.allowed_schema_ids(),
+                         [])
+
+