+++ ebpub/ebpub/utils/view_utils.py
@@ -85,7 +85,50 @@ def has_staff_cookie(request):
     return request.COOKIES.get(settings.STAFF_COOKIE_NAME) == settings.STAFF_COOKIE_VALUE
 
 def get_schema_manager(request):
+    """
+    Returns a Manager that restricts the Schemas that can be seen
+    based on the current request.
+
+    This should be used in ALL public view queries that reference
+    NewsItems or Schemas.  There should probably be a more convenient
+    API for doing this.
+
+    By default, this just uses ``has_staff_cookie`` to decide whether
+    to show Schemas that are not public, but you can also name the
+    path to a function in settings.SCHEMA_MANAGER_HOOK
+    that has a signature like func(manager, request.user)
+    and returns a manager.
+    """
     if has_staff_cookie(request):
-        return Schema.objects
+        manager = Schema.objects
     else:
-        return Schema.public_objects
+        manager = Schema.public_objects
+    return _manager_filter_hook(manager, request.user)
+
+def get_schema_manager_for_user(user):
+    """Like get_schema_manager(request),
+    but useful in contexts where you have a User object available
+    but no request, eg. batch jobs like sending email.
+
+    This can also be overridden/enhanced by setting
+    settings.SCHEMA_MANAGER_HOOK.
+    """
+    if user.is_superuser:
+        manager = Schema.objects
+    else:
+        manager = Schema.public_objects
+    return _manager_filter_hook(manager, user)
+
+def _manager_filter_hook(manager, user):
+    # Hook to customize the Manager's behavior based on the current user.
+    # The named function should take (user, manager) arguments
+    # and return something that behaves like a manager but can do whatever
+    # extra filtering you like in eg. get_query_set().
+    mgr_filter = getattr(settings, 'SCHEMA_MANAGER_HOOK', None)
+    if mgr_filter is not None:
+        module, func = mgr_filter.split(':')
+        import importlib
+        module = importlib.import_module(module)
+        func = getattr(module, func)
+        manager = func(user, manager)
+    return manager